music_player/requirement.txt
mutagen==1.47.0
pillow==11.0.0
pygame==2.6.1
